rfct: Pkg comment on main, telegram; Renamed e to util
